refactor(keyboard): hoist duplicated input-range helpers
input.ts and blockEdit.ts declared identical InputTargetRange /
RawSelectionFailureReason types and identical
rawRangeFromInputEvent / rawRangeFromTargetRange /
rawSelectionReason helpers. Move them into inputRange.ts and
import the single public entry from both files.
