+++ packages/core/src/features/keyboard/blockEdit.ts
@@ -1,6 +1,6 @@
 import {htmlChildren, isHtmlElement} from '../../shared/checkers'
 import {KEYBOARD} from '../../shared/constants'
-import type {BoundaryPositionResult, Range, RawSelectionResult} from '../../shared/editorContracts'
+import type {Range} from '../../shared/editorContracts'
 import {listen} from '../../shared/signals/index.js'
 import type {Store} from '../../store/Store'
 
@@ -10,15 +10,7 @@ import {addDragRow, getMergeDragRowJoinPos, mergeDragRows, canMergeRows} from '.
 import {consumeMarkupPaste} from '../clipboard'
 import type {Token} from '../parsing'
 import * as caretDom from '../selection/caretDom'
-
-type InputTargetRange = {
-	readonly startContainer: Node
-	readonly startOffset: number
-	readonly endContainer: Node
-	readonly endOffset: number
-}
-
-type RawSelectionFailureReason = Extract<RawSelectionResult, {ok: false}>['reason']
+import {rawRangeFromInputEvent} from './inputRange'
 
 function isTextLikeRow(token: Token): boolean {
 	if (token.type === 'text') return true
@@ -324,32 +316,6 @@ function replaceBlockRange(store: KbCtx, event: InputEvent, replacement: string)
 
 	event.preventDefault()
 	store.edit.replace(range, replacement)
-}
-
-function rawRangeFromInputEvent(store: KbCtx, event: InputEvent): RawSelectionResult {
-	const ranges = event.getTargetRanges()
-	if (ranges.length === 0) return store.dom.readRawSelection()
-	return rawRangeFromTargetRange(store, ranges[0])
-}
-
-function rawRangeFromTargetRange(store: KbCtx, range: InputTargetRange): RawSelectionResult {
-	const start = store.dom.rawPositionFromBoundary(range.startContainer, range.startOffset, 'after')
-	const end = store.dom.rawPositionFromBoundary(range.endContainer, range.endOffset, 'before')
-	if (!start.ok) return {ok: false, reason: rawSelectionReason(start)}
-	if (!end.ok) return {ok: false, reason: rawSelectionReason(end)}
-	return {
-		ok: true,
-		value: {
-			range:
-				start.value <= end.value ? {start: start.value, end: end.value} : {start: end.value, end: start.value},
-		},
-	}
-}
-
-function rawSelectionReason(result: BoundaryPositionResult): RawSelectionFailureReason {
-	if (result.ok) return 'invalidBoundary'
-	if (result.reason === 'composing') return 'invalidBoundary'
-	return result.reason
 }
 
 function rangeForBlockInput(store: KbCtx, event: InputEvent, range: Range): Range | undefined {
